Prevent layout shifting from vertical scrollbar
The layout root container width is set to the viewport width
Horizontal overflow is set to hidden

With longer pages, the appearing vertical scrollbar causes the root
container to squish horizontally a little amount, shifting some of the
components inside. Setting the width to viewport width (100vw) keeps
the container size consistent and responds only to window size change.

As a side effect, a horizontal scrollbar appears. Horizontal overflow
is set to hidden to disable scrolling horizontally.