Collapse top navigation on smaller screens
This ensures that the page scales correctly on smaller devices.

As the screen size gets smaller, navigation links are hidden.
The links are then contained in a hamburger menu.

Clicking the button in the top right will open the menu.
When menu is open, clicking the links or anywhere outside will close it.

Header component keeps the state of the open menu.

For the moment, the styling for active page is disabled.